cricknova_ai_backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
numpy
opencv-python-headless
razorpay
google-generativeai>=0.8.0
paypal-checkout-serversdk==1.0.3
google-cloud-firestore>=2.16.0
